HDDS-6497. Return ozone.ico
