+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/server/http/HttpServer2.java
@@ -803,7 +803,6 @@ public final class HttpServer2 implements FilterContainer {
     addServlet("logLevel", "/logLevel", LogLevel.Servlet.class);
     addServlet("jmx", "/jmx", JMXJsonServlet.class);
     addServlet("conf", "/conf", ConfServlet.class);
-    addServlet("icon", "/favicon.ico", IconServlet.class);
   }
 
   public void addContext(ServletContextHandler ctxt, boolean isFiltered) {
@@ -1517,18 +1516,6 @@ public final class HttpServer2 implements FilterContainer {
     return adminsAcl != null && adminsAcl.isUserAllowed(remoteUserUGI);
   }
 
-  /**
-   * Do nothing for favicon.ico.
-   */
-  public static class IconServlet extends HttpServlet {
-    private static final long serialVersionUID = -1L;
-
-    @Override
-    public void doGet(HttpServletRequest request, HttpServletResponse response)
-        throws ServletException, IOException {
-      return;
-    }
-  }
 
   /**
    * A very simple servlet to serve up a text representation of the current
